refactor: SzppIcon.vueをatoms/へ
