+++ src/components/atoms/SzppIcon.vue
